requirements.txt
# Core FastAPI Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
sqlalchemy>=2.0.0
pydantic>=2.0.0

# Authentication and Security
passlib[bcrypt]>=1.7.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.9

# Google APIs and OAuth
google-auth>=2.40.0
google-api-python-client>=2.170.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
google-ai-generativelanguage>=0.4.0

# Vector Database and AI
faiss-cpu>=1.11.0
numpy>=2.0.0
google-cloud-aiplatform>=1.45.0

# Data Validation and Serialization
email-validator>=2.0.0

# HTTP and Networking
requests>=2.31.0
httplib2>=0.22.0
urllib3>=2.0.0

# Testing
pytest>=8.0.0

# Utilities
typing-extensions>=4.0.0
anyio>=4.0.0
sniffio>=1.0.0